Moved rich text utilities to a helper class
